Update interface to reflect change to ms

+++ src/Framework/EventLoop/Interfaces/SchedulerInterface.php
@@ -7,8 +7,8 @@ interface SchedulerInterface
 {
     public function task(Closure $callback): void;
     public function defer(Closure $closure): void;
-    public function interval(float $interval, Closure $callback);
-    public function delay(float $delay, Closure $callback);
+    public function interval(int $interval, Closure $callback);
+    public function delay(int $delay, Closure $callback);
     public function io($resource, ?Closure $callback);
 
     public function attach($resource, ?Closure $onRead = null, ?Closure $onWrite = null): bool;
